# REF/PERF: ArrowExtensionArray.__setitem__