--- a/pandas/core/arrays/arrow/array.py
+++ b/pandas/core/arrays/arrow/array.py
@@ -589,7 +589,23 @@ class ArrowExtensionArray(OpsMixin, ExtensionArray):
                     f" expected {len(self)}"
                 )
 
-        fill_value = self._maybe_convert_setitem_value(value)
+        def convert_fill_value(value, pa_type, dtype):
+            if value is None:
+                return value
+            if isinstance(value, (pa.Scalar, pa.Array, pa.ChunkedArray)):
+                return value
+            if is_array_like(value):
+                pa_box = pa.array
+            else:
+                pa_box = pa.scalar
+            try:
+                value = pa_box(value, type=pa_type, from_pandas=True)
+            except pa.ArrowTypeError as err:
+                msg = f"Invalid value '{str(value)}' for dtype {dtype}"
+                raise TypeError(msg) from err
+            return value
+
+        fill_value = convert_fill_value(value, self._data.type, self.dtype)
 
         try:
             if method is None:
@@ -1029,7 +1045,9 @@ class ArrowExtensionArray(OpsMixin, ExtensionArray):
             # fast path
             key = cast(int, key)
             n = len(self)
-            if not (-n <= key < n):
+            if key < 0:
+                key += n
+            if not 0 <= key < n:
                 raise IndexError(
                     f"index {key} is out of bounds for axis 0 with size {n}"
                 )