@jbrockmendel 2023-01-09
Not for this PR but because I'm cranky: the fact that we have to implement a kludgy `__setitem__` sucks.  ideally pyarrow would implement it; failing that better to set into each of the elements of `self._data.buffers()` (which besides being a PITA im not sure is even possible in mask-less cases)