@lukemanley 2023-01-10
> Not for this PR but because I'm cranky: the fact that we have to implement a kludgy `__setitem__` sucks. ideally pyarrow would implement it; failing that better to set into each of the elements of `self._data.buffers()` (which besides being a PITA im not sure is even possible in mask-less cases)

This was a (failed?) attempt to make it a little less kludgy :). Agreed its still kludgy. 

Part of the complexity is supporting older versions of pyarrow where the pyarrow.compute functions have bugs we need to work around. We could remove the fast paths if simplicity is preferred.

I'm not familiar with the buffers but could take a look.